refactor: LadderStatistics를 LadderResults로 수정
