+++ src/main/java/ladder/controller/LadderController.java
@@ -1,6 +1,7 @@
 package ladder.controller;
 
 import ladder.domain.*;
+import ladder.domain.engine.LadderResults;
 import ladder.view.InputView;
 import ladder.view.ResultView;
 
@@ -14,7 +15,7 @@ public class LadderController {
     private Players players;
     private ExecutionResults executionResults;
     private Ladder ladder;
-    private LadderStatistics ladderStatistics;
+    private LadderResults ladderResults;
     private final static Map<Point, String> printPoints = new HashMap<>();
     private final static String ALL_PLAYERS = "all";
 
@@ -58,11 +59,11 @@ public class LadderController {
         String playerName = InputView.enterPlayerYouWant();
         Map<String, String> results;
         if (playerName.equalsIgnoreCase(ALL_PLAYERS)) {
-            results = ladderStatistics.resultsOfAll();
+            results = ladderResults.resultsOfAll();
             ResultView.printExecutionResult(results);
             return;
         }
-        results = ladderStatistics.results(new Player(playerName));
+        results = ladderResults.results(new Player(playerName));
         ResultView.printExecutionResult(results);
     }
 
@@ -71,7 +72,7 @@ public class LadderController {
         printLadder();
         printResult();
 
-        ladderStatistics = ladder.ladderStatistics(players, executionResults);
+        ladderResults = ladder.ladderResults(players, executionResults);
         printExecutionResult();
         printExecutionResult();
     }
